# Conflicts:
#	docs/acknowledgements.txt
#	scripts/linux/install-daemon.sh
#	src/Raven.Studio/typescript/viewmodels/database/advanced/databaseIDs.ts
#	src/Raven.Studio/wwwroot/App/views/database/advanced/databaseIDs.html
#	src/Raven.Studio/wwwroot/App/views/database/status/statistics.html
#	src/Raven.Studio/wwwroot/App/views/database/tasks/ongoingTasks.html
#	test/Tryouts/Program.cs

+++ src/Corax/IndexEntry/IndexEntryReader.cs
@@ -320,7 +320,7 @@ public unsafe readonly ref struct IndexEntryReader
                 if (PtrBitVector.GetBitInPointer(nullTablePtr + spanTableOffset, elementIdx) == true)
                     goto HasNull;
 
-                int nullBitStreamSize = totalElements / (sizeof(long) * 8) + totalElements % (sizeof(long) * 8) == 0 ? 0 : 1;
+                int nullBitStreamSize = totalElements / (sizeof(byte) * 8) + (totalElements % (sizeof(byte) * 8) == 0 ? 0 : 1);
                 spanTableOffset += nullBitStreamSize; // Point after the null table.                             
             }
